Summarizing inferences based on cross-validation

Summary method for cv.ncvreg objects

# S3 method for cv.ncvreg
summary(object, ...)
# S3 method for summary.cv.ncvreg
print(x, digits, ...)

Arguments

object
A "cv.ncvreg" object.

x
A "summary.cv.ncvreg" object.

digits
Number of digits past the decimal point to print out. Can be a vector specifying different display digits for each of the five non-integer printed values.

…
Further arguments passed to or from other methods.

Value

summary.cv.ncvreg produces an object with S3 class "summary.cv.ncvreg". The class has its own print method and contains the following list elements:

penalty

The penalty used by ncvreg.

model

Either "linear" or "logistic", depending on the family option in ncvreg.

n

Number of observations

p

Number of regression coefficients (not including the intercept).

min

The index of lambda with the smallest cross-validation error.

lambda

The sequence of lambda values used by cv.ncvreg.

cve

Cross-validation error (deviance).

r.squared

Proportion of variance explained by the model, as estimated by cross-validation.

snr

Signal to noise ratio, as estimated by cross-validation.

sigma

For linear regression models, the scale parameter estimate.

pe

For logistic regression models, the prediction error (misclassification error).

References

Breheny, P. and Huang, J. (2011) Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection. Ann. Appl. Statist., 5: 232-253.

See also

Examples

# Linear regression --------------------------------------------------
data(Prostate)
cvfit <- cv.ncvreg(Prostate$X, Prostate$y)
summary(cvfit)

#> MCP-penalized linear regression with n=97, p=8
#> At minimum cross-validation error (lambda=0.0008):
#> -------------------------------------------------
#>   Nonzero coefficients: 8
#>   Cross-validation error (deviance): 0.55
#>   R-squared: 0.58
#>   Signal-to-noise ratio: 1.38
#>   Scale estimate (sigma): 0.744

# Logistic regression ------------------------------------------------
data(Heart)
cvfit <- cv.ncvreg(Heart$X, Heart$y, family="binomial")
summary(cvfit)

#> MCP-penalized logistic regression with n=462, p=9
#> At minimum cross-validation error (lambda=0.0102):
#> -------------------------------------------------
#>   Nonzero coefficients: 7
#>   Cross-validation error (deviance): 1.06
#>   R-squared: 0.18
#>   Signal-to-noise ratio: 0.22
#>   Prediction error: 0.266