Summarizing cross-validation-based inference

Summary method for cv.ncvreg objects

# S3 method for class 'cv.ncvreg'
summary(object, ...)

# S3 method for class 'summary.cv.ncvreg'
print(x, digits, ...)

Arguments

object

A cv.ncvreg or cv.ncvsurv object.

...

Further arguments passed to or from other methods.

x

A summary.cv.ncvreg object.

digits

Number of digits past the decimal point to print out. Can be a vector specifying different display digits for each of the five non-integer printed values.

Value

An object with S3 class summary.cv.ncvreg. The class has its own print method and contains the following list elements:

penalty

The penalty used by ncvreg.

model

Either "linear" or "logistic", depending on the family option in ncvreg.

n

Number of instances

p

Number of regression coefficients (not including the intercept).

min

The index of lambda with the smallest cross-validation error.

lambda

The sequence of lambda values used by cv.ncvreg.

cve

Cross-validation error (deviance).

r.squared

Proportion of variance explained by the model, as estimated by cross-validation. For models outside of linear regression, the Cox-Snell approach to defining R-squared is used.

snr

Signal to noise ratio, as estimated by cross-validation.

sigma

For linear regression models, the scale parameter estimate.

pe

For logistic regression models, the prediction error (misclassification error).

References

Breheny P and Huang J. (2011) Coordinate descent algorithms for nonconvex penalized regression, with applications to biological feature selection. Annals of Applied Statistics, 5: 232-253. doi:10.1214/10-AOAS388

See also

Author

Patrick Breheny

Examples

# Linear regression --------------------------------------------------
data(Prostate)
cvfit <- cv.ncvreg(Prostate$X, Prostate$y)
summary(cvfit)
#> MCP-penalized linear regression with n=97, p=8
#> At minimum cross-validation error (lambda=0.0638):
#> -------------------------------------------------
#>   Nonzero coefficients: 5
#>   Cross-validation error (deviance): 0.53
#>   R-squared: 0.60
#>   Signal-to-noise ratio: 1.51
#>   Scale estimate (sigma): 0.725
#> MCP-penalized linear regression with n=97, p=8
#> At lambda=0.0638:
#> -------------------------------------------------
#>   Nonzero coefficients         :   5
#>   Expected nonzero coefficients:   1.87
#>   Average mfdr (5 features)    :   0.374
#> 
#>          Estimate      z     mfdr Selected
#> lcavol   0.533903  8.500  < 1e-04        *
#> svi      0.692676  3.873 0.015225        *
#> lweight  0.611121  3.537 0.050561        *
#> lbph     0.054944  1.584 0.885443        *
#> age     -0.006941 -1.331 0.917808        *

# Logistic regression ------------------------------------------------
data(Heart)
cvfit <- cv.ncvreg(Heart$X, Heart$y, family="binomial")
summary(cvfit)
#> MCP-penalized logistic regression with n=462, p=9
#> At minimum cross-validation error (lambda=0.0177):
#> -------------------------------------------------
#>   Nonzero coefficients: 7
#>   Cross-validation error (deviance): 1.07
#>   R-squared: 0.20
#>   Signal-to-noise ratio: 0.25
#>   Prediction error: 0.262
#> MCP-penalized logistic regression with n=462, p=9
#> At lambda=0.0177:
#> -------------------------------------------------
#>   Nonzero coefficients         :   7
#>   Expected nonzero coefficients:   1.86
#>   Average mfdr (7 features)    :   0.266
#> 
#>          Estimate      z      mfdr Selected
#> age      0.050443  5.812   < 1e-04        *
#> famhist  0.911639  4.124 0.0023397        *
#> tobacco  0.080099  3.301 0.0451490        *
#> typea    0.037675  3.208 0.0601656        *
#> ldl      0.170154  3.160 0.0693819        *
#> obesity -0.015959 -1.322 0.8199744        *
#> sbp      0.001333  1.054 0.8623941        *

# Cox regression -----------------------------------------------------
data(Lung)
cvfit <- cv.ncvsurv(Lung$X, Lung$y)
summary(cvfit)
#> MCP-penalized Cox regression with n=137, p=8
#> At minimum cross-validation error (lambda=0.1276):
#> -------------------------------------------------
#>   Nonzero coefficients: 3
#>   Cross-validation error (deviance): 7.50
#>   R-squared: 0.31
#>   Signal-to-noise ratio: 0.46
#> MCP-penalized Cox regression with n=137, p=8
#> At lambda=0.1276:
#> -------------------------------------------------
#>   Nonzero coefficients         :   3
#>   Expected nonzero coefficients:   1.28
#>   Average mfdr (3 features)    :   0.426
#> 
#>          Estimate      z    mfdr Selected
#> karno    -0.03318 -6.589 < 1e-04        *
#> squamous -0.38295 -2.732 0.59462        *
#> adeno     0.33542  2.585 0.68452        *